# docs/swagger.yaml
definitions:
  APIError:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  DeleteMessageReq:
    properties:
      messageId:
        type: string
      phone:
        type: string
    required:
    - messageId
    - phone
    type: object
...
  EditMessageReq:
    properties:
      body:
        type: string
      messageId:
        type: string
      phone:
        type: string
    required:
    - body
    - messageId
    - phone
    type: object
...
  MidResp:
    properties:
      messageId:
        type: string
    type: object
...
  SendButtonReq:
    properties:
      body:
        type: string
      buttons:
        items:
          $ref: '#/definitions/ButtonItem'
        minItems: 1
        type: array
      customId:
        type: string
      footer:
        type: string
      mentionedJids:
        items:
          type: string
        type: array
      phone:
        type: string
      replyTo:
        $ref: '#/definitions/ReplyContext'
    required:
    - body
    - buttons
    - phone
    type: object
  SendContactReq:
    properties:
      name:
        type: string
      phone:
        type: string
      vcard:
        type: string
    required:
    - name
    - phone
    - vcard
    type: object
...
  SendMediaReq:
    properties:
      base64:
        type: string
      caption:
        type: string
      customId:
        type: string
      fileName:
        type: string
      mentionedJids:
        items:
          type: string
        type: array
      mimeType:
        type: string
      phone:
        type: string
      replyTo:
        $ref: '#/definitions/ReplyContext'
      url:
        type: string
    required:
    - mimeType
    - phone
    type: object
...
info:
  contact:
    name: wzap Support
    url: https://github.com/felipyfgs/wzap
  description: WhatsApp Multi-Session API powered by whatsmeow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: wzap API
  version: "1.0"
paths:
...
  /sessions/{sessionId}/messages/audio:
    post:
      consumes:
      - application/json
      description: Sends a base64-encoded audio file as a voice note (PTT)
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Media payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SendMediaReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/MidResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Send an audio message
      tags:
      - Messages
  /sessions/{sessionId}/messages/button:
    post:
      consumes:
      - application/json
      description: Sends a message with interactive buttons
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Button message payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SendButtonReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/MidResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Send a button message
      tags:
      - Messages
  /sessions/{sessionId}/messages/contact:
    post:
      consumes:
      - application/json
      description: Sends a vCard contact message via WhatsApp to the specified recipient
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Contact payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SendContactReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/MidResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Send a contact card
      tags:
      - Messages
  /sessions/{sessionId}/messages/delete:
    post:
      consumes:
      - application/json
      description: Revokes a previously sent message for all recipients (unsend)
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Delete payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/DeleteMessageReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/MidResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Delete a sent message
      tags:
      - Messages
  /sessions/{sessionId}/messages/document:
    post:
      consumes:
      - application/json
      description: Sends a base64-encoded file as a document. Use filename to set
        the display name.
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Media payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SendMediaReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/MidResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Send a document
      tags:
      - Messages
  /sessions/{sessionId}/messages/edit:
    post:
      consumes:
      - application/json
      description: Edits a previously sent message by mid, replacing its text content
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Edit payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/EditMessageReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/MidResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Edit a sent message
      tags:
      - Messages
  /sessions/{sessionId}/messages/forward:
    post:
      consumes:
      - application/json
      description: Forwards an existing message to another chat or group
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Forward message payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ForwardMessageReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/MidResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Forward a message
      tags:
      - Messages
  /sessions/{sessionId}/messages/image:
    post:
      consumes:
      - application/json
      description: Sends a base64-encoded image to a WhatsApp JID with optional caption
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Media payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SendMediaReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/MidResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Send an image message
      tags:
      - Messages
  /sessions/{sessionId}/messages/link:
    post:
      consumes:
      - application/json
      description: Sends a hyperlink with optional title and description as a rich
        preview message
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Link payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SendLinkReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/MidResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Send a link preview message
      tags:
      - Messages
  /sessions/{sessionId}/messages/list:
    post:
      consumes:
      - application/json
      description: Sends a message with interactive list sections
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: List message payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SendListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/MidResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Send a list message
      tags:
      - Messages
  /sessions/{sessionId}/messages/location:
    post:
      consumes:
      - application/json
      description: Sends a GPS location message with optional name and address to
        the specified recipient
      parameters:
      - description: Session name or ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Location payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SendLocationReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/APIResponse'
            - properties:
                Data:
                  $ref: '#/definitions/MidResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      security:
      - Authorization: []
      summary: Send a location message
      tags:
      - Messages
...
securityDefinitions:
  Authorization:
    description: Session or admin token (without Bearer prefix)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"